* CC: use timer delay to show documentation; prevents scroll stuttering
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@8919 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/codecompletion/doxygen_parser.cpp
@@ -557,7 +557,7 @@ wxString DocumentationHelper::DoxygenToHTML(const wxString& doc)
         keyword = parser.FindNextKeyword(doc);
     }
     // process nested keywords:
-    for (int i = 0; i < (size_t)(sizeof(arguments)/sizeof(arguments[0])); ++i)
+    for (size_t i = 0; i < (size_t)(sizeof(arguments)/sizeof(arguments[0])); ++i)
     {
         arguments[i].Trim(true).Trim(false);
